zvision-wa/hermes-agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
boto3>=1.35.0
botocore>=1.35.0
pypdf>=5.0.0
pydantic>=2.12.0
python-dotenv>=1.2.0
aiofiles>=24.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
